refactor: API 参照と frame-rpc を共有パッケージへ分離
- レンダラー向け API と frame-rpc を共有パッケージとして切り出し
- apps/api を追加し、api.ts と window.api の型定義を @repo/api へ移設
- packages/frame-rpc を追加し、apps/web の各参照を workspace package import に置換
- apps/web の依存関係と pnpm-lock.yaml を更新し、旧 your-app-name 依存を削除
