+++ apps/api/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@repo/api",
+  "version": "0.2.0-alpha.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/api.ts",
+      "import": "./src/api.ts"
+    }
+  },
+  "scripts": {
+    "lint": "tsgo --noEmit && oxlint",
+    "check": "tsgo --noEmit && oxlint --fix"
+  },
+  "dependencies": {
+    "@repo/frame-rpc": "workspace:*",
+    "your-app-name": "workspace:*"
+  },
+  "devDependencies": {
+    "@types/node": "^24.12.2",
+    "@typescript/native-preview": "7.0.0-dev.20260418.1",
+    "oxlint": "^1.60.0",
+    "typescript": "^6.0.3"
+  },
+  "packageManager": "pnpm@10.33.0"
+}
