+++ apps/web/package.json
@@ -14,6 +14,7 @@
     "@faker-js/faker": "^10.3.0",
     "@repo/ai-chat": "workspace:*",
     "@repo/ai-tools": "workspace:*",
+    "@repo/api": "workspace:*",
     "@repo/shadcn": "workspace:*",
     "@standard-schema/spec": "^1.1.0",
     "@tailwindcss/vite": "^4.2.2",
@@ -35,7 +36,6 @@
     "react-markdown": "^10.1.0",
     "remark-gfm": "^4.0.1",
     "tailwindcss": "^4.1.18",
-    "your-app-name": "workspace:*",
     "zod": "^4.3.6"
   },
   "devDependencies": {
