+++ apps/web/src/routes/(app)/demo.kakeibo.tsx
@@ -16,9 +16,9 @@ import {
 import type { Table as TanStackTable } from '@tanstack/react-table'
 import { Square, Table } from 'lucide-react'
 
+import { kakeibo } from '@repo/api'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@repo/shadcn/ui/tabs'
 
-import { kakeibo } from '@/api'
 import { BasicTable } from '@/components/table/basic-table'
 import { fuzzyFilter } from '@/components/table/tableUtils'
