+++ apps/web/src/routes/(app)/ui/index.tsx
@@ -4,6 +4,7 @@ import { createFileRoute, useNavigate } from '@tanstack/react-router'
 import { MoonIcon, SunIcon } from 'lucide-react'
 import { z } from 'zod'
 
+import type { ThemeApi } from '@repo/api'
 import { components } from '@repo/shadcn/demo/constants'
 import { cn } from '@repo/shadcn/lib/utils'
 import { Button } from '@repo/shadcn/ui/button'
@@ -30,7 +31,6 @@ import {
 } from '@repo/shadcn/ui/popover'
 import { ToggleGroup, ToggleGroupItem } from '@repo/shadcn/ui/toggle-group'
 
-import type { ThemeApi } from '@/api'
 import { useTheme as useAppearanceMode } from '@/components/theme-provider'
 import { ThemeSwitcher as ModeSwitcher } from '@/components/theme-switcher'
 import { useStyle } from '@/features/style/api/use-style'
